refactor: rename `/internationalization` → `/i18n`
