+++ studio/studioConfig.tsx
@@ -6,7 +6,7 @@ import { structureTool } from "sanity/structure";
 import { internationalizedArray } from "sanity-plugin-internationalized-array";
 import { media } from "sanity-plugin-media";
 
-import { languageID } from "internationalization/languageSchemaField";
+import { languageID } from "i18n/languageSchemaField";
 
 import StudioIcon from "./components/studioIcon/StudioIcon";
 import { deskStructure } from "./deskStructure";
